+++ forage-android/src/main/java/com/joinforage/forage/android/core/services/forageapi/paymentmethod/FetchPaymentMethodService.kt
@@ -3,22 +3,16 @@ package com.joinforage.forage.android.core.services.forageapi.paymentmethod
 import com.joinforage.forage.android.core.services.ForageConfig
 import com.joinforage.forage.android.core.services.forageapi.engine.IHttpEngine
 import com.joinforage.forage.android.core.services.forageapi.requests.GetPaymentMethodRequest
-import com.joinforage.forage.android.ecom.services.forageapi.requests.PostEcomPaymentMethodRequest
 
-internal class PaymentMethodResponse(val json: String) {
-    val parsed = PaymentMethod(json)
-}
-
-internal interface IPaymentMethodService {
+internal interface IFetchPaymentMethodService {
     suspend fun fetchPaymentMethod(paymentMethodRef: String): PaymentMethodResponse
-    suspend fun createPaymentMethod(rawPan: String, customerId: String?, reusable: Boolean): PaymentMethodResponse
 }
 
-internal open class PaymentMethodService(
+internal open class FetchPaymentMethodService(
     protected val forageConfig: ForageConfig,
     protected val traceId: String,
     protected val engine: IHttpEngine
-) : IPaymentMethodService {
+) : IFetchPaymentMethodService {
 
     class FailedToFetchPaymentMethodException(val paymentMethodRef: String, e: Exception) :
         Exception("Failed to fetch payment method $paymentMethodRef", e)
@@ -34,25 +28,8 @@ internal open class PaymentMethodService(
     } catch (e: Exception) {
         throw FailedToFetchPaymentMethodException(paymentMethodRef, e)
     }
-
-    class FailedToCreatePaymentMethodException(e: Exception) :
-        Exception("Failed to create payment method", e)
-
-    override suspend fun createPaymentMethod(
-        rawPan: String,
-        customerId: String?,
-        reusable: Boolean
-    ): PaymentMethodResponse = try {
-        engine.sendRequest(
-            PostEcomPaymentMethodRequest(
-                forageConfig,
-                traceId,
-                rawPan,
-                customerId,
-                reusable
-            )
-        ).let { PaymentMethodResponse(it) }
-    } catch (e: Exception) {
-        throw FailedToCreatePaymentMethodException(e)
-    }
 }
+
+
+
+
