+++ forage-android/src/main/java/com/joinforage/forage/android/core/services/forageapi/paymentmethod/PaymentMethodResponse.kt
@@ -1,0 +1,5 @@
+package com.joinforage.forage.android.core.services.forageapi.paymentmethod
+
+internal class PaymentMethodResponse(val json: String) {
+    val parsed = PaymentMethod(json)
+}
