+++ forage-android/src/main/java/com/joinforage/forage/android/ecom/services/vault/submission/EcomBalanceCheckSubmission.kt
@@ -2,7 +2,6 @@ package com.joinforage.forage.android.ecom.services.vault.submission
 
 import com.joinforage.forage.android.core.services.ForageConfig
 import com.joinforage.forage.android.core.services.forageapi.network.ForageApiResponse
-import com.joinforage.forage.android.core.services.forageapi.paymentmethod.IPaymentMethodService
 import com.joinforage.forage.android.core.services.forageapi.requests.ClientApiRequest
 import com.joinforage.forage.android.core.services.forageapi.requests.RosettaVaultRequest
 import com.joinforage.forage.android.core.services.telemetry.LogLogger
@@ -16,6 +15,7 @@ import com.joinforage.forage.android.core.services.vault.requests.ISubmitRequest
 import com.joinforage.forage.android.core.services.vault.submission.BalanceCheckSubmission
 import com.joinforage.forage.android.core.services.vault.submission.ISubmitDelegate
 import com.joinforage.forage.android.core.services.vault.submission.PinSubmission
+import com.joinforage.forage.android.ecom.services.forageapi.paymentmethod.IPaymentMethodService
 import com.joinforage.forage.android.ecom.services.vault.EcomBaseBodyBuilder
 
 private class EcomRosettaBalanceInquiryRequest(
