+++ forage-android/src/main/java/com/joinforage/forage/android/ecom/services/vault/submission/EcomCapturePaymentSubmission.kt
@@ -3,7 +3,6 @@ package com.joinforage.forage.android.ecom.services.vault.submission
 import com.joinforage.forage.android.core.services.ForageConfig
 import com.joinforage.forage.android.core.services.forageapi.network.ForageApiResponse
 import com.joinforage.forage.android.core.services.forageapi.payment.IPaymentService
-import com.joinforage.forage.android.core.services.forageapi.paymentmethod.IPaymentMethodService
 import com.joinforage.forage.android.core.services.forageapi.requests.ClientApiRequest
 import com.joinforage.forage.android.core.services.telemetry.LogLogger
 import com.joinforage.forage.android.core.services.telemetry.UserAction
@@ -16,6 +15,7 @@ import com.joinforage.forage.android.core.services.vault.metrics.VaultMetricsRec
 import com.joinforage.forage.android.core.services.vault.requests.ISubmitRequestBuilder
 import com.joinforage.forage.android.core.services.vault.requests.RosettaCapturePaymentRequest
 import com.joinforage.forage.android.core.services.vault.submission.PinSubmission
+import com.joinforage.forage.android.ecom.services.forageapi.paymentmethod.IPaymentMethodService
 import com.joinforage.forage.android.ecom.services.vault.EcomBaseBodyBuilder
 
 private class EcomRosettaCapturePaymentRequest(
